+++ src/main/java/com/example/AccesaProject/repository/QuestRepository.java
@@ -3,5 +3,5 @@ package com.example.AccesaProject.repository;
 import com.example.AccesaProject.entity.Quest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestRepository extends JpaRepository<Quest,Integer> {
+public interface QuestRepository extends JpaRepository<Quest,Long> {
 }
